Give each of databaseTests and blackboxTests its own entry point
