+++ src/blackboxTests/src/main/kotlin/org/vaccineimpact/api/blackboxTests/App.kt
@@ -1,0 +1,8 @@
+package org.vaccineimpact.api.blackboxTests
+
+import org.vaccineimpact.api.test_helpers.DatabaseCreationHelper
+
+fun main(args: Array<String>)
+{
+    DatabaseCreationHelper.handleEntryPoint(args)
+}
