+++ src/databaseTests/src/main/kotlin/org/vaccineimpact/api/databaseTests/App.kt
@@ -1,0 +1,8 @@
+package org.vaccineimpact.api.databaseTests
+
+import org.vaccineimpact.api.test_helpers.DatabaseCreationHelper
+
+fun main(args: Array<String>)
+{
+    DatabaseCreationHelper.handleEntryPoint(args)
+}
