+++ src/databaseTests/src/main/resources/config.properties
@@ -1,0 +1,1 @@
+../../test/resources/config.properties
