+++ src/testHelpers/build.gradle
@@ -1,6 +1,3 @@
-apply plugin: 'application'
-mainClassName = "org.vaccineimpact.api.test_helpers.AppKt"
-
 dependencies {
     compile "junit:junit:4.12"
     compile "org.jooq:jooq:3.10.5"
@@ -10,18 +7,4 @@ dependencies {
     compile project(':databaseInterface')
     compile project(':models')
     compile project(':security')
-}
-
-processResources.finalizedBy 'copyConfig'
-
-task createTemplateFromDatabase(type: JavaExec) {
-    main = "org.vaccineimpact.api.test_helpers.AppKt"
-    classpath = sourceSets.main.runtimeClasspath
-    args "createTemplateFromDatabase"
-}
-
-task restoreDatabaseFromTemplate(type: JavaExec) {
-    main = "org.vaccineimpact.api.test_helpers.AppKt"
-    classpath = sourceSets.main.runtimeClasspath
-    args "restoreDatabaseFromTemplate"
 }
