--- src/testHelpers/src/main/resources/config.properties
@@ -1,16 +1,0 @@
-db.host=${db_host}
-db.port=${db_port}
-db.name=${db_name}
-db.username=${db_username}
-db.password=${db_password}
-
-annex.host=${annex_host}
-annex.port=${annex_port}
-annex.name=${annex_name}
-annex.username=${annex_username}
-annex.password=${annex_password}
-
-testdb.template_name=${testdb_template_name}
-annex.template_name=${annex_template_name}
-
-app.url=${app_url}
